requirements.txt
firecrawl-py>=1.4.0
playwright>=1.44.0
requests>=2.31.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-json-report>=1.5.0
python-dotenv>=1.0.0
